Add files via upload
This game will be made using the windows form .net in visual studio.  In this game you can jump using the space bar, with each jump player will switch gravity and jump from bottom platform to the top platform. We are using animated GIF images for the player, there is two GIF image one for each platform landing. 
To make the game more challenging we’ve added two obstacles that gets positioned randomly in the game and will move towards the player, this will make the play jump between the two platforms and last as long as possible in the game. The game will keep score and a high score. End of the game you will be able to see how much you score and if it gets saved to the high score value.